moving noticeOfWorkActionCreator tests and NoW mocks to common dir
