+++ services/core-web/src/tests/actionCreators/documentActionCreator.spec.js
@@ -5,7 +5,7 @@ import { ENVIRONMENT } from "@mds/common/constants/environment";
 import * as COMMON_API from "@mds/common/constants/API";
 import * as API from "@/constants/API";
 import * as MOCK from "@mds/common/tests/mocks/dataMocks";
-import * as NOW_MOCK from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOW_MOCK from "@mds/common/tests/mocks/noticeOfWorkMock";
 import {
   exportNoticeOfWorkApplicationDocument,
   fetchExplosivesPermitDocumentContextTemplate,
