+++ services/core-web/src/tests/components/Forms/noticeOfWork/EditNoWContacts.spec.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { EditNoWContacts } from "@/components/Forms/noticeOfWork/EditNoWContacts";
-import * as NOW_MOCK from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOW_MOCK from "@mds/common/tests/mocks/noticeOfWorkMock";
 
 const dispatchProps = {};
 const props = {};
