+++ services/core-web/src/tests/components/modalContent/UpdateStatusGenerateLetterModal.spec.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { UpdateStatusGenerateLetterModal } from "@/components/modalContent/UpdateStatusGenerateLetterModal";
-import * as NOW_MOCK from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOW_MOCK from "@mds/common/tests/mocks/noticeOfWorkMock";
 
 const dispatchProps = {};
 const props = {};
