+++ services/core-web/src/tests/components/noticeOfWork/applications/FinalPermitDocuments.spec.js
@@ -2,7 +2,7 @@ import React from "react";
 import { shallow } from "enzyme";
 import { FinalPermitDocuments } from "@/components/noticeOfWork/applications/FinalPermitDocuments";
 import * as MOCK from "@mds/common/tests/mocks/dataMocks";
-import { IMPORTED_NOTICE_OF_WORK } from "@/tests/mocks/noticeOfWorkMocks";
+import { IMPORTED_NOTICE_OF_WORK } from "@mds/common/tests/mocks/noticeOfWorkMock";
 
 const dispatchProps = {};
 const props = {};
