+++ services/core-web/src/tests/components/noticeOfWork/applications/NOWReasonForDelay.spec.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { NOWReasonForDelay } from "@/components/noticeOfWork/applications/NOWReasonForDelay";
-import * as NOWMocks from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOWMocks from "@mds/common/tests/mocks/noticeOfWorkMock";
 
 const reducerProps = {};
