+++ services/core-web/src/tests/components/noticeOfWork/applications/NoticeOfWorkApplication.spec.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { NoticeOfWorkApplication } from "@/components/noticeOfWork/applications/NoticeOfWorkApplication";
-import * as NOW_MOCK from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOW_MOCK from "@mds/common/tests/mocks/noticeOfWorkMock";
 
 const dispatchProps: any = {};
 const reducerProps: any = {};
