+++ services/core-web/src/tests/components/noticeOfWork/applications/PreviousAmendmentDocuments.spec.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { PermitPackage } from "@/components/noticeOfWork/applications/PermitPackage";
-import * as NOWMocks from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOWMocks from "@mds/common/tests/mocks/noticeOfWorkMock";
 
 const props = {};
 const dispatchProps = {};
 
-const setupDispatchProps = () => {};
+const setupDispatchProps = () => { };
 
 const setupProps = () => {
   props.editPreambleFileMetadata = false;
