+++ services/core-web/src/tests/components/noticeOfWork/applications/ScrollContentWrapper.spec.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { ScrollContentWrapper } from "@/components/noticeOfWork/applications/ScrollContentWrapper";
-import * as NOWMocks from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOWMocks from "@mds/common/tests/mocks/noticeOfWorkMock";
 
 const props = {};
 const dispatchProps = {};
