+++ services/core-web/src/tests/components/noticeOfWork/applications/review/activities/WaterSupply.spec.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { WaterSupply } from "@/components/noticeOfWork/applications/review/activities/WaterSupply";
-import * as NOW_MOCK from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOW_MOCK from "@mds/common/tests/mocks/noticeOfWorkMock";
 
 const reducerProps = {};
