+++ services/core-web/src/tests/components/noticeOfWork/applications/verification/VerificationTab.spec.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { VerificationTab } from "@/components/noticeOfWork/applications/verification/VerificationTab";
-import * as NOW_MOCK from "@/tests/mocks/noticeOfWorkMocks";
+import * as NOW_MOCK from "@mds/common/tests/mocks/noticeOfWorkMock";
 import * as MOCK from "@mds/common/tests/mocks/dataMocks";
 
 const dispatchProps = {};
