--- services/core-web/src/tests/mocks/noticeOfWorkMocks.js
@@ -1,2671 +1,0 @@
-export const NOTICE_OF_WORK = {
-  application_guid: "89ce70bc-a34b-4ec6-aff0-704ba68c4606",
-  now_application_guid: "07e801a0-fa33-4c3b-abcc-ac6df628d483",
-  mine_guid: "20af050d-2a1a-4870-9016-eacabaa68f02",
-  mine_name: null,
-  mine_region: "NW",
-  trackingnumber: null,
-  applicationtype: null,
-  status: null,
-  submitteddate: null,
-  receiveddate: null,
-  applicantclientid: 42,
-  submitterclientid: 43,
-  noticeofworktype: null,
-  typeofpermit: null,
-  typeofapplication: null,
-  minenumber: null,
-  latitude: null,
-  longitude: null,
-  nameofproperty: null,
-  tenurenumbers: null,
-  crowngrantlotnumbers: null,
-  sitedirections: null,
-  firstaidequipmentonsite: null,
-  firstaidcertlevel: null,
-  descexplorationprogram: null,
-  proposedstartdate: null,
-  proposedenddate: null,
-  yearroundseasonal: null,
-  landcommunitywatershed: null,
-  landprivate: null,
-  landlegaldesc: null,
-  archsitesaffected: null,
-  isonprivateland: null,
-  sandgravelquarryoperations: null,
-  storeexplosivesonsite: null,
-  bcexplosivespermitissued: null,
-  bcexplosivespermitnumber: null,
-  bcexplosivespermitexpiry: null,
-  campdisturbedarea: null,
-  camptimbervolume: null,
-  bldgdisturbedarea: null,
-  bldgtimbervolume: null,
-  stgedisturbedarea: null,
-  stgetimbervolume: null,
-  fuellubstoreonsite: null,
-  fuellubstored: null,
-  fuellubstoremethodbulk: null,
-  fuellubstoremethodbarrel: null,
-  cbsfreclamation: null,
-  cbsfreclamationcost: null,
-  mechtrenchingreclamation: null,
-  mechtrenchingreclamationcost: null,
-  expsurfacedrillreclamation: null,
-  expsurfacedrillreclcorestorage: null,
-  expsurfacedrillreclamationcost: null,
-  expaccessreclamation: null,
-  expaccessreclamationcost: null,
-  surfacebulksampleprocmethods: null,
-  surfacebulksamplereclamation: null,
-  surfacebulksamplereclsephandl: null,
-  surfacebulksamplerecldrainmiti: null,
-  surfacebulksamplereclcost: null,
-  underexptotalore: null,
-  underexptotaloreunits: null,
-  underexptotalwaste: null,
-  underexptotalwasteunits: null,
-  underexpreclamation: null,
-  underexpreclamationcost: null,
-  placerundergroundoperations: null,
-  placerhandoperations: null,
-  placerreclamationarea: null,
-  placerreclamation: null,
-  placerreclamationcost: null,
-  sandgrvqrydepthoverburden: null,
-  sandgrvqrydepthtopsoil: null,
-  sandgrvqrystabilizemeasures: null,
-  sandgrvqrywithinaglandres: null,
-  sandgrvqryalrpermitnumber: null,
-  sandgrvqrylocalgovsoilrembylaw: null,
-  sandgrvqryofficialcommplan: null,
-  sandgrvqrylandusezoning: null,
-  sandgrvqryendlanduse: null,
-  sandgrvqrytotalmineres: null,
-  sandgrvqrytotalmineresunits: null,
-  sandgrvqryannualextrest: null,
-  sandgrvqryannualextrestunits: null,
-  sandgrvqryreclamation: null,
-  sandgrvqryreclamationbackfill: null,
-  sandgrvqryreclamationcost: null,
-  sandgrvqrygrdwtravgdepth: null,
-  sandgrvqrygrdwtrexistingareas: null,
-  sandgrvqrygrdwtrtestpits: null,
-  sandgrvqrygrdwtrtestwells: null,
-  sandgrvqrygrdwtrother: null,
-  sandgrvqrygrdwtrmeasprotect: null,
-  sandgrvqryimpactdistres: null,
-  sandgrvqryimpactdistwater: null,
-  sandgrvqryimpactnoise: null,
-  sandgrvqryimpactprvtaccess: null,
-  sandgrvqryimpactprevtdust: null,
-  sandgrvqryimpactminvisual: null,
-  cutlinesexplgridtotallinekms: null,
-  cutlinesexplgridtimbervolume: null,
-  cutlinesreclamation: null,
-  cutlinesreclamationcost: null,
-  pondswastewatertreatfacility: null,
-  freeusepermit: null,
-  licencetocut: null,
-  timbertotalvolume: null,
-  campbuildstgetotaldistarea: null,
-  mechtrenchingtotaldistarea: null,
-  expsurfacedrilltotaldistarea: null,
-  expaccesstotaldistarea: null,
-  surfacebulksampletotaldistarea: null,
-  placertotaldistarea: null,
-  underexptotaldistarea: null,
-  sandgrvqrytotaldistarea: null,
-  pondstotaldistarea: null,
-  reclcostsubtotal: null,
-  reclcostexist: null,
-  reclcostrecl: null,
-  reclcosttotal: null,
-  reclareasubtotal: null,
-  reclareaexist: null,
-  reclarearecl: null,
-  reclareatotal: null,
-  application_reason_codes: [],
-  anyotherinformation: null,
-  vfcbcapplicationurl: null,
-  messagecreateddate: null,
-  processed: null,
-  processeddate: null,
-  cutlinesexplgriddisturbedarea: null,
-  pondsrecycled: null,
-  pondsexfiltratedtoground: null,
-  pondsdischargedtoenv: null,
-  pondsreclamation: null,
-  pondsreclamationcost: null,
-  sandgrvqrytotalexistdistarea: null,
-  nrsosapplicationid: null,
-  isblastselect: null,
-  istimberselect: null,
-  applicant: {
-    type: null,
-    org_legalname: null,
-    org_doingbusinessas: null,
-    ind_firstname: null,
-    ind_lastname: null,
-    ind_middlename: null,
-    ind_phonenumber: null,
-    dayphonenumber: null,
-    dayphonenumberext: null,
-    faxnumber: null,
-    email: null,
-    org_bcfedincorpnumber: null,
-    org_bcregnumber: null,
-    org_societynumber: null,
-    org_hstregnumber: null,
-    org_contactname: null,
-    mailingaddressline1: null,
-    mailingaddressline2: null,
-    mailingaddresscity: null,
-    mailingaddressprovstate: null,
-    mailingaddresscountry: null,
-    mailingaddresspostalzip: null,
-  },
-  submitter: {
-    type: null,
-    org_legalname: null,
-    org_doingbusinessas: null,
-    ind_firstname: null,
-    ind_lastname: null,
-    ind_middlename: null,
-    ind_phonenumber: null,
-    dayphonenumber: null,
-    dayphonenumberext: null,
-    faxnumber: null,
-    email: "someone@submitter.com",
-    org_bcfedincorpnumber: null,
-    org_bcregnumber: null,
-    org_societynumber: null,
-    org_hstregnumber: null,
-    org_contactname: null,
-    mailingaddressline1: null,
-    mailingaddressline2: null,
-    mailingaddresscity: null,
-    mailingaddressprovstate: null,
-    mailingaddresscountry: null,
-    mailingaddresspostalzip: null,
-  },
-  documents: [
-    {
-      mine_document_guid: "12345-678",
-      documenturl: null,
-      filename: null,
-      documenttype: null,
-      description: null,
-    },
-    {
-      mine_document_guid: "12345-6789",
-      documenturl: null,
-      filename: null,
-      documenttype: null,
-      description: null,
-    },
-  ],
-  filtered_submission_documents: [
-    {
-      messageid: 0,
-      documenturl: null,
-      description: null,
-      document_manager_guid: null,
-      documenttype: null,
-      filename: null,
-      is_final_package: false,
-      mine_document_guid: null,
-      now_application_document_xref_guid: null,
-      now_application_id: null,
-    },
-  ],
-  contacts: [
-    {
-      type: null,
-      org_legalname: null,
-      org_doingbusinessas: null,
-      ind_firstname: null,
-      ind_lastname: null,
-      ind_middlename: null,
-      ind_phonenumber: null,
-      dayphonenumber: null,
-      dayphonenumberext: null,
-      faxnumber: null,
-      email: null,
-      org_bcfedincorpnumber: null,
-      org_bcregnumber: null,
-      org_societynumber: null,
-      org_hstregnumber: null,
-      org_contactname: null,
-      mailingaddressline1: null,
-      contacttype: null,
-      contactcertificationtype: null,
-      contactcertificationid: null,
-      mailingaddressline2: null,
-      mailingaddresscity: null,
-      mailingaddressprovstate: null,
-      mailingaddresscountry: null,
-      mailingaddresspostalzip: null,
-      seq_no: null,
-    },
-    {
-      type: null,
-      org_legalname: null,
-      org_doingbusinessas: null,
-      ind_firstname: null,
-      ind_lastname: null,
-      ind_middlename: null,
-      ind_phonenumber: null,
-      dayphonenumber: null,
-      dayphonenumberext: null,
-      faxnumber: null,
-      email: null,
-      org_bcfedincorpnumber: null,
-      org_bcregnumber: null,
-      org_societynumber: null,
-      org_hstregnumber: null,
-      org_contactname: null,
-      mailingaddressline1: null,
-      contacttype: null,
-      contactcertificationtype: null,
-      contactcertificationid: null,
-      mailingaddressline2: null,
-      mailingaddresscity: null,
-      mailingaddressprovstate: null,
-      mailingaddresscountry: null,
-      mailingaddresspostalzip: null,
-      seq_no: null,
-    },
-  ],
-  existing_placer_activity: [
-    {
-      type: null,
-      quantity: null,
-      depth: null,
-      length: null,
-      width: null,
-      disturbedarea: null,
-      timbervolume: null,
-    },
-  ],
-  existing_settling_pond: [
-    {
-      pondid: null,
-      watersource: null,
-      width: null,
-      length: null,
-      depth: null,
-      constructionmethod: null,
-      disturbedarea: null,
-      timbervolume: null,
-    },
-  ],
-  proposed_placer_activity: [
-    {
-      type: null,
-      quantity: null,
-      depth: null,
-      length: null,
-      width: null,
-      disturbedarea: null,
-      timbervolume: null,
-    },
-  ],
-  proposed_settling_pond: [
-    {
-      pondid: null,
-      watersource: null,
-      width: null,
-      length: null,
-      depth: null,
-      constructionmethod: null,
-      disturbedarea: null,
-      timbervolume: null,
-    },
-  ],
-  surface_bulk_sample_activity: [
-    {
-      type: null,
-      disturbedarea: 1.11,
-      timbervolume: null,
-      quantity: null,
-    },
-  ],
-  sand_grv_qry_activity: [
-    {
-      type: null,
-      disturbedarea: null,
-      timbervolume: null,
-    },
-  ],
-  under_exp_new_activity: [
-    {
-      type: null,
-      incline: null,
-      inclineunits: null,
-      quantity: 400,
-      length: null,
-      width: null,
-      height: null,
-      seq_no: null,
-    },
-  ],
-  under_exp_rehab_activity: [
-    {
-      type: null,
-      incline: null,
-      inclineunits: null,
-      quantity: 711,
-      length: null,
-      width: null,
-      height: null,
-      seq_no: null,
-    },
-  ],
-  under_exp_surface_activity: [
-    {
-      type: null,
-      quantity: 810,
-      disturbedarea: null,
-      timbervolume: null,
-    },
-  ],
-  water_source_activity: [
-    {
-      sourcewatersupply: null,
-      type: null,
-      useofwater: null,
-      estimateratewater: null,
-      pumpsizeinwater: null,
-      locationwaterintake: null,
-      seq_no: null,
-    },
-  ],
-  exp_access_activity: [],
-  exp_surface_drill_activity: [],
-  mech_trenching_activity: [],
-  site_property: {
-    mine_commodity_code: [],
-    mine_disturbance_code: [],
-    mine_tenure_type_code: "",
-  },
-};
-
-export const IMPORT_NOTICE_OF_WORK_SUBMISSION_DOCUMENTS_JOB = {
-  import_now_submission_documents_job_id: 17,
-  start_timestamp: "2020-11-17T17:20:13.327639",
-  end_timestamp: "2020-11-17T17:20:13.340830",
-  create_timestamp: "2020-11-17T17:16:54.375057",
-  complete_timestamp: "2020-11-17T17:20:13.340830",
-  attempt: 1,
-  create_user: null,
-  celery_task_id: "da2c1d64-0328-4674-810c-609ecefa1e61",
-  now_application_id: "1",
-  now_application_guid: "89ce70bc-a34b-4ec6-aff0-704ba68c4606",
-  import_now_submission_documents_job_status_code: "SUC",
-  import_now_submission_documents: [
-    {
-      import_now_submission_documents_job_id: 17,
-      document_id: 660,
-      submission_document_url: "https://",
-      submission_document_file_name: "Multi-year Exploration Area.dbf",
-      submission_document_type: "SpatialFileDoc",
-      submission_document_message_id: 0,
-      error: null,
-    },
-  ],
-};
-
-export const NOTICE_OF_WORK_LIST = {
-  records: [NOTICE_OF_WORK],
-  current_page: 1,
-  items_per_page: 5,
-  total: 25,
-  total_pages: 5,
-};
-
-export const IMPORTED_NOTICE_OF_WORK = {
-  now_application_id: "",
-  now_application_guid: "1346wesrt22",
-  mine_guid: "",
-  now_message_id: 47284,
-  now_tracking_number: 47284,
-  notice_of_work_type_code: "MIN",
-  submitted_date: "",
-  received_date: "",
-  proposed_start_date: "2016-03-01",
-  proposed_end_date: "2020-03-01",
-  adjusted_annual_maximum_tonnage: "",
-  proposed_annual_maximum_tonnage: 10000,
-  latitude: "",
-  longitude: "",
-  property_name: "",
-  tenure_number: "",
-  description_of_land: "",
-  contacts: [],
-  blasting_operation: {},
-  state_of_land: {},
-  camp: { details: [] },
-  cut_lines_polarization_survey: { details: [] },
-  exploration_access: { details: [] },
-  exploration_surface_drilling: { details: [] },
-  mechanical_trenching: { details: [] },
-  sand_gravel_quarry_operation: { details: [] },
-  settling_pond: { details: [] },
-  surface_bulk_sample: { details: [] },
-  underground_exploration: { details: [] },
-  water_supply: { details: [] },
-  placer_operation: { details: [] },
-  documents: [],
-  submission_documents: [],
-  filtered_submission_documents: [],
-  application_documents: [],
-  lead_inspector: {},
-};
-
-export const NOTICE_OF_WORK_ACTIVITY_TYPES = {
-  records: [
-    {
-      activity_type_code: "cut_lines_polarization_survey",
-      description: "Cut Lines and Induced Polarization Survey",
-    },
-    { activity_type_code: "water_supply", description: "Water Supply" },
-  ],
-};
-
-export const DROPDOWN_NOTICE_OF_WORK_ACTIVITY_TYPES = [
-  {
-    value: "cut_lines_polarization_survey",
-    label: "Cut Lines and Induced Polarization Survey",
-    subType: null,
-    isActive: true,
-  },
-  { value: "water_supply", label: "Water Supply", subType: null, isActive: true },
-];
-
-export const ACTIVITY_TYPE_HASH = {
-  cut_lines_polarization_survey: "Cut Lines and Induced Polarization Survey",
-  water_supply: "Water Supply",
-};
-
-export const RECLAMATION_SUMMARY = [
-  {
-    label: "Cut Lines and Induced Polarization Survey",
-    total: "0.00",
-    cost: "0.00",
-  },
-  {
-    label: "Water Supply",
-    total: "0.00",
-    cost: "0.00",
-  },
-];
-
-export const EQUIPMENT = [
-  {
-    capacity: "John Deere 690",
-    description: "Excavator",
-    quantity: 1,
-  },
-];
-
-export const UNDERGROUND_EXPLORATION_TYPES = {
-  records: [
-    {
-      underground_exploration_type_code: "NEW",
-      description: "New",
-    },
-    {
-      underground_exploration_type_code: "RHB",
-      description: "Rehabilitation",
-    },
-    {
-      underground_exploration_type_code: "SUR",
-      description: "Surface",
-    },
-  ],
-};
-
-export const DROPDOWN_UNDERGROUND_EXPLORATION_TYPES = [
-  {
-    value: "NEW",
-    label: "New",
-    subType: null,
-    isActive: true,
-  },
-  { value: "RHB", label: "Rehabilitation", subType: null, isActive: true },
-  { value: "SUR", label: "Surface", subType: null, isActive: true },
-];
-
-export const UNDERGROUND_EXPLORATION_TYPES_HASH = {
-  NEW: "New",
-  RHB: "Rehabilitation",
-  SUR: "Surface",
-};
-
-export const APPLICATION_TYPES = {
-  records: [
-    {
-      notice_of_work_type_code: "QCA",
-      description: "Quarry - Construction Aggregate",
-    },
-    {
-      notice_of_work_type_code: "COL",
-      description: "Coal",
-    },
-    {
-      notice_of_work_type_code: "PLA",
-      description: "Placer Operations",
-    },
-    {
-      notice_of_work_type_code: "MIN",
-      description: "Mineral",
-    },
-    {
-      notice_of_work_type_code: "SAG",
-      description: "Sand & Gravel",
-    },
-    {
-      notice_of_work_type_code: "QIM",
-      description: "Quarry - Industrial Mineral",
-    },
-  ],
-};
-
-export const DROPDOWN_APPLICATION_TYPES = [
-  { value: "COL", label: "Coal", subType: null, isActive: true },
-  {
-    value: "MIN",
-    label: "Mineral",
-    subType: null,
-    isActive: true,
-  },
-  { value: "PLA", label: "Placer Operations", subType: null, isActive: true },
-  {
-    value: "QCA",
-    label: "Quarry - Construction Aggregate",
-    subType: null,
-    isActive: true,
-  },
-  { value: "QIM", label: "Quarry - Industrial Mineral", subType: null, isActive: true },
-  { value: "SAG", label: "Sand & Gravel", subType: null, isActive: true },
-];
-
-export const APPLICATION_TYPES_HASH = {
-  QCA: "Quarry - Construction Aggregate",
-  COL: "Coal",
-  PLA: "Placer Operations",
-  MIN: "Mineral",
-  SAG: "Sand & Gravel",
-  QIM: "Quarry - Industrial Mineral",
-};
-
-export const APPLICATION_STATUS_CODES = {
-  records: [
-    { now_application_status_code: "SUB", description: "Submitted", active_ind: true },
-    { now_application_status_code: "REF", description: "Referred", active_ind: true },
-    { now_application_status_code: "CDI", description: "Client Delay Info", active_ind: true },
-    { now_application_status_code: "CDB", description: "Client Delay Bond", active_ind: true },
-    { now_application_status_code: "GVD", description: "Govt Delay", active_ind: true },
-    { now_application_status_code: "CON", description: "Consultation", active_ind: true },
-    { now_application_status_code: "AIA", description: "Active/Issued/Approved", active_ind: true },
-    { now_application_status_code: "WDN", description: "Withdrawn", active_ind: true },
-    { now_application_status_code: "REJ", description: "Rejected", active_ind: true },
-    { now_application_status_code: "CLO", description: "Closed", active_ind: true },
-  ],
-};
-
-export const DROPDOWN_APPLICATION_STATUS_CODES = [
-  { value: "AIA", label: "Active/Issued/Approved", subType: null, isActive: true },
-  { value: "CDB", label: "Client Delay Bond", subType: null, isActive: true },
-  { value: "CDI", label: "Client Delay Info", subType: null, isActive: true },
-  { value: "CLO", label: "Closed", subType: null, isActive: true },
-  { value: "CON", label: "Consultation", subType: null, isActive: true },
-  { value: "GVD", label: "Govt Delay", subType: null, isActive: true },
-  { value: "REF", label: "Referred", subType: null, isActive: true },
-  { value: "REJ", label: "Rejected", subType: null, isActive: true },
-  {
-    value: "SUB",
-    label: "Submitted",
-    subType: null,
-    isActive: true,
-  },
-  { value: "WDN", label: "Withdrawn", subType: null, isActive: true },
-];
-
-export const APPLICATION_STATUS_CODES_HASH = {
-  SUB: "Submitted",
-  REF: "Referred",
-  CDI: "Client Delay Info",
-  CDB: "Client Delay Bond",
-  GVD: "Govt Delay",
-  CON: "Consultation",
-  AIA: "Active/Issued/Approved",
-  WDN: "Withdrawn",
-  REJ: "Rejected",
-  CLO: "Closed",
-};
-
-export const UNIT_TYPES = {
-  records: [
-    {
-      unit_type_code: "MTN",
-      short_description: "Tonne (Metric Ton 1,000 kg)",
-      active_ind: true,
-    },
-    {
-      unit_type_code: "MEC",
-      short_description: "Meters cubed",
-      active_ind: true,
-    },
-    {
-      unit_type_code: "HA",
-      short_description: "Hectares",
-      active_ind: true,
-    },
-    {
-      unit_type_code: "DEG",
-      short_description: "Degrees",
-      active_ind: true,
-    },
-    {
-      unit_type_code: "PER",
-      short_description: "Grade (Percent)",
-      active_ind: true,
-    },
-    {
-      unit_type_code: "MTR",
-      short_description: "Meters",
-      active_ind: true,
-    },
-  ],
-};
-
-export const DROPDOWN_UNIT_TYPES = [
-  {
-    label: "%",
-    value: "PER",
-    subType: null,
-    isActive: true,
-  },
-  {
-    label: "deg",
-    value: "DEG",
-    subType: null,
-    isActive: true,
-  },
-  {
-    label: "ha",
-    value: "HA",
-    subType: null,
-    isActive: true,
-  },
-  {
-    label: "km",
-    value: "KMT",
-    subType: null,
-    isActive: true,
-  },
-  {
-    label: "m",
-    value: "MTR",
-    subType: null,
-    isActive: true,
-  },
-  {
-    label: "m³",
-    value: "MEC",
-    subType: null,
-    isActive: true,
-  },
-  {
-    label: "t",
-    value: "MTN",
-    subType: null,
-    isActive: true,
-  },
-];
-
-export const UNIT_TYPES_HASH = {
-  KMT: "km",
-  MTN: "t",
-  MEC: "m³",
-  HA: "ha",
-  DEG: "deg",
-  PER: "%",
-  MTR: "m",
-};
-
-export const APPLICATION_DOCUMENT_TYPES = [
-  {
-    now_application_document_type_code: "PFR",
-    description: "Preliminary Field Reconnaissance",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "CAL",
-    description: "Acknowledgement Letter",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {
-      document_template_code: "NCL",
-      form_spec: [
-        {
-          id: "letter_dt",
-          label: "Letter Date",
-          type: "DATE",
-          placeholder: "YYYY-MM-DD",
-          required: true,
-          "context-value": "{DATETIME.UTCNOW}",
-          "read-only": false,
-        },
-        {
-          id: "mine_no",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "proponent_name",
-          label: "Proponent Name",
-          type: "FIELD",
-          placeholder: "Enter the proponent's name",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "proponent_address",
-          label: "Proponent Address",
-          type: "AUTO_SIZE_FIELD",
-          placeholder: "Enter the proponent's address",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "emailed_to",
-          label: "Emailed to",
-          type: "FIELD",
-          placeholder: "Enter the name of the email recipient",
-          required: false,
-          "read-only": false,
-        },
-        {
-          id: "property",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_dt_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "This letter acknowledges receipt of your Notice of Work and Reclamation Program dated",
-          "read-only": false,
-        },
-        {
-          id: "application_dt",
-          label: "Application Date",
-          type: "DATE",
-          placeholder: "YYYY-MM-DD",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "exploration_type_label",
-          type: "LABEL",
-          required: false,
-          "context-value": "Your proposed program of",
-          "read-only": false,
-        },
-        {
-          id: "exploration_type",
-          label: "Exploration Type",
-          type: "FIELD",
-          placeholder: "Enter the exploration type",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "bond_inc_amt_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "has been referred to other resource agencies and has been sent to Indigenous Nations for consultation. Prior to the approval and issuance of your permit, you are required to post a security deposit of",
-          "read-only": false,
-        },
-        {
-          id: "bond_inc_amt",
-          label: "Bond Increase Amount",
-          type: "CURRENCY",
-          placeholder: "Enter the bond increase amount",
-          required: false,
-          "read-only": false,
-        },
-        {
-          id: "letter_body_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "and you may wish to take the opportunity to post your security at this time to avoid any delays.  Safekeeping Agreements backed by GIC’s may be used for bonds under $25,000 with the enclosed template. Complete the form with your banker, using the 'Instructions on Completing a Safekeeping Agreement' and return it to this office for our signature.  A copy of the completed form will be returned to you and your financial institution.  Irrevocable Letters of Credit, certified cheque, bank draft or money order made payable to the Minister of Finance, at the undernoted address, are also acceptable. Payments made by EFT can also be arranged. Please do not send cash. For reclamation surety bonds, the bond shall be with a surety licensed to transact the business of a surety in Canada. For the surety bond template and more please visit our Reclamation Security website. Personal information collected by the Ministry of Mining and Critical Minerals is under the authority of section 26(c) of the Freedom of Information and Protection of Privacy Act for the purpose of collecting Bond and Securities Data. If you have any questions about the collection, use and disclosure of your personal information, please contact: Mines Digital Services by email at mds@gov.bc.ca, by phone at: 778-698-7233, or by mail at: PO Box 9380, STN PROV GOVT, Victoria, BC, V8W 9M6.",
-          "read-only": false,
-        },
-        {
-          id: "letter_body",
-          label: "Letter Body",
-          type: "AUTO_SIZE_FIELD",
-          required: true,
-          "context-value":
-            "Other legislation may be applicable to the operation and you (the Permittee) may be required to obtain approvals or permits under that legislation. It is your responsibility to comply with the terms and conditions of all other permits and authorizations which you may have been issued and other applicable legislation including, but not limited to the: Wildlife Act, Wildfire Act, Wildfire Regulation and the Water Sustainability Act.",
-          "read-only": false,
-        },
-        {
-          id: "issuing_inspector_name_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "You are reminded that no work may commence until you have received your permits. To clarify or discuss any of the above, please call or email me at the information below.\n\nSincerely,",
-          "read-only": false,
-        },
-        {
-          id: "issuing_inspector_name",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_email",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_phone",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_email",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_phone_number",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_fax_number",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_mailing_address_line_1",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_mailing_address_line_2",
-          required: false,
-          "read-only": true,
-        },
-      ],
-    },
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "WDL",
-    description: "Withdrawal Letter",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {
-      document_template_code: "NWL",
-      form_spec: [
-        {
-          id: "letter_dt",
-          label: "Letter Date",
-          type: "DATE",
-          placeholder: "YYYY-MM-DD",
-          required: true,
-          "context-value": "{DATETIME.UTCNOW}",
-          "read-only": false,
-        },
-        {
-          id: "mine_no",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "proponent_name",
-          label: "Proponent Name",
-          type: "FIELD",
-          placeholder: "Enter the proponent's name",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "proponent_address",
-          label: "Proponent Address",
-          type: "AUTO_SIZE_FIELD",
-          placeholder: "Enter the proponent's address",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "property",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "withdrawal_dt_label",
-          type: "LABEL",
-          required: false,
-          "context-value": "I refer to your decision of",
-          "read-only": false,
-        },
-        {
-          id: "withdrawal_dt",
-          label: "Withdrawal Date",
-          type: "DATE",
-          placeholder: "YYYY-MM-DD",
-          required: true,
-          "context-value": "{DATETIME.UTCNOW}",
-          "read-only": false,
-        },
-        {
-          id: "letter_body_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "to withdraw your Notice of Work application and confirm that all further processing of your application has now been terminated.",
-          "read-only": false,
-        },
-        {
-          id: "letter_body",
-          label: "Letter Body",
-          type: "AUTO_SIZE_FIELD",
-          required: true,
-          "context-value":
-            "You will have to reapply should you wish to carry out your intended work program. You are reminded that pursuant to Section 10 of the Mines Act no exploration activities can be carried out unless you have received the required permit.",
-          "read-only": false,
-        },
-        {
-          id: "issuing_inspector_name_label",
-          type: "LABEL",
-          required: false,
-          "context-value": "Sincerely,",
-          "read-only": false,
-        },
-        {
-          id: "issuing_inspector_name",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_email",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_phone",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_email",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_phone_number",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_fax_number",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_mailing_address_line_1",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_mailing_address_line_2",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_type_code",
-          required: false,
-          "read-only": true,
-        },
-      ],
-    },
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "REV",
-    description: "Review",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "ECC",
-    description: "Email Correspondence/Communications",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "NPR",
-    description: "No Permit Required",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "LMA",
-    description: "Location Map",
-    now_application_document_sub_type_code: "MDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "NPI",
-    description: "No Permit Required IP",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "OMA",
-    description: "Overview Map",
-    now_application_document_sub_type_code: "MDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SMA",
-    description: "Supplemental Map",
-    now_application_document_sub_type_code: "MDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RFD",
-    description: "Reasons for Decision",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SOP",
-    description: "Standard/Safe Operating Procedures",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RSP",
-    description: "Riparian Setbacks Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "WMP",
-    description: "Water Management Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "WPL",
-    description: "Wildlife Management Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RPL",
-    description: "Reclamation Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "OMP",
-    description: "Other Management Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SEP",
-    description: "Sediment and Erosion Control Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "FDP",
-    description: "Fugitive Dust Management Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "VMP",
-    description: "Vegetation Management Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "TSS",
-    description: "Terrain Stability Study",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "MAD",
-    description: "Metal Leaching/Acid Rock Drainage",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "LNO",
-    description: "Landowner Notification",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "DWP",
-    description: "Description of Work/Work Program",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "ARE",
-    description: "Agent Letter of Representation",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "AMR",
-    description: "Amendment Request",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "MYA",
-    description: "MYAB Update Form",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SUD",
-    description: "Supporting Documents",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SRE",
-    description: "Status Report",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SOM",
-    description: "Status Report - Overlapping Interests Maps",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SRS",
-    description: "Status Report - Shape Files",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "WFI",
-    description: "30 day Warning for Information",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "WFS",
-    description: "30 day Warning for Security",
-    now_application_document_sub_type_code: "SDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "MPW",
-    description: "Proposed and/or Permitted Mine Area Map",
-    now_application_document_sub_type_code: "MDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "CRS",
-    description: "Consultation Report/Summary",
-    now_application_document_sub_type_code: "CDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "TMP",
-    description: "Title/Tenure Map",
-    now_application_document_sub_type_code: "MDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "BCR",
-    description: "Begin Consultation Request",
-    now_application_document_sub_type_code: "CDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SCD",
-    description: "Bond Calculator",
-    now_application_document_sub_type_code: "SDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SRB",
-    description: "Scan of Reclamation Security Document",
-    now_application_document_sub_type_code: "SDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "NIA",
-    description: "No Interest Acknowledgement Form",
-    now_application_document_sub_type_code: "SDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "AKL",
-    description: "Acknowledgement of Security Letter",
-    now_application_document_sub_type_code: "SDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "OTH",
-    description: "Other",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "CCC",
-    description: "Consultation Correspondence (not in CRTS)",
-    now_application_document_sub_type_code: "CDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "CSD",
-    description: "Consultation Support for Decision",
-    now_application_document_sub_type_code: "CDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "BRR",
-    description: "Begin Referral Request",
-    now_application_document_sub_type_code: "RDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RSR",
-    description: "Referral Summary Roll Up",
-    now_application_document_sub_type_code: "RDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RLE",
-    description: "Referral Letter (outside of E-Referral)",
-    now_application_document_sub_type_code: "RDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RRE",
-    description: "Referral Response (outside of E-Referral)",
-    now_application_document_sub_type_code: "RDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "PCA",
-    description: "Advertisement",
-    now_application_document_sub_type_code: "PDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "PCC",
-    description: "Public Comment",
-    now_application_document_sub_type_code: "PDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "PCM",
-    description: "Ministry Response",
-    now_application_document_sub_type_code: "PDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RMI",
-    description: "Request for More Information",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "LTM",
-    description: "Land Title/Licence of Occupation Map",
-    now_application_document_sub_type_code: "MDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "OTA",
-    description: "Other",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "MPG",
-    description: "Proposed and/or Permitted Mine Area Map",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RJL",
-    description: "Rejection Letter",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {
-      document_template_code: "NRL",
-      form_spec: [
-        {
-          id: "letter_dt",
-          label: "Letter Date",
-          type: "DATE",
-          placeholder: "YYYY-MM-DD",
-          required: true,
-          "context-value": "{DATETIME.UTCNOW}",
-          "read-only": false,
-        },
-        {
-          id: "mine_no",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "proponent_name",
-          label: "Proponent Name",
-          type: "FIELD",
-          placeholder: "Enter the proponent's name",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "proponent_address",
-          label: "Proponent Address",
-          type: "AUTO_SIZE_FIELD",
-          placeholder: "Enter the proponent's address",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "property",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_dt_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "This letter serves as formal notice that the Notice of Work and Reclamation application dated ",
-          "read-only": false,
-        },
-        {
-          id: "application_dt",
-          label: "Application Date",
-          type: "DATE",
-          placeholder: "YYYY-MM-DD",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "letter_body_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "for the above noted property has been discontinued for the proposed project.",
-          "read-only": false,
-        },
-        {
-          id: "letter_body",
-          label: "Letter Body",
-          type: "AUTO_SIZE_FIELD",
-          required: true,
-          "context-value":
-            "Future proposals for mining activities on the above noted property will require the submission of a new Notice of Work application. Should you require further information or have questions please do not hesitate to contact me.",
-          "read-only": false,
-        },
-        {
-          id: "issuing_inspector_name_label",
-          type: "LABEL",
-          required: false,
-          "context-value": "Sincerely,",
-          "read-only": false,
-        },
-        {
-          id: "issuing_inspector_name",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_email",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_phone",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_email",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_phone_number",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_fax_number",
-          required: true,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_mailing_address_line_1",
-          required: true,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_mailing_address_line_2",
-          required: true,
-          "read-only": true,
-        },
-        {
-          id: "application_type_code",
-          required: false,
-          "read-only": true,
-        },
-      ],
-    },
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "NPE",
-    description: "Permit Enclosed Letter",
-    now_application_document_sub_type_code: "GDO",
-    document_template: {
-      document_template_code: "NPE",
-      form_spec: [
-        {
-          id: "letter_dt",
-          label: "Letter Date",
-          type: "DATE",
-          placeholder: "YYYY-MM-DD",
-          required: true,
-          "context-value": "{DATETIME.UTCNOW}",
-          "read-only": false,
-        },
-        {
-          id: "mine_no",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "proponent_name",
-          label: "Proponent Name",
-          type: "FIELD",
-          placeholder: "Enter the proponent's name",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "proponent_address",
-          label: "Proponent Address",
-          type: "AUTO_SIZE_FIELD",
-          placeholder: "Enter the proponent's address",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "property",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_dt_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "Please find enclosed your Mines Act permit, which authorizes exploration activities as detailed in the Notice of Work and Reclamation Program dated",
-          "read-only": false,
-        },
-        {
-          id: "application_dt",
-          label: "Application Date",
-          type: "DATE",
-          placeholder: "YYYY-MM-DD",
-          required: true,
-          "read-only": false,
-        },
-        {
-          id: "letter_body_label",
-          type: "LABEL",
-          required: false,
-          "context-value":
-            "The Notice of Work and Reclamation Program form part of your permit, and you are reminded that you may not depart from the permitted program without written authorization.",
-          "read-only": false,
-        },
-        {
-          id: "letter_body",
-          label: "Letter Body",
-          type: "AUTO_SIZE_FIELD",
-          required: true,
-          "context-value":
-            "Please ensure that you and all persons who are carrying out activities in accordance with this permit comply with all terms and conditions of the permit and are familiar with the permitted work program.\n\nThis permit applies only to the requirements under the Mines Act and Health, Safety and Reclamation Code for Mines in British Columbia (Code).  Other legislation may be applicable to the operation and you (the Permittee) may be required to obtain approvals or permits under that legislation. Examples of other authorizations would be for timber removal, water use, works within the agricultural land reserve etc.\n\nThe amount of your security deposit may be adjusted on the basis of reclamation performance, field inspections by this ministry, and on reports which may be requested.",
-          "read-only": false,
-        },
-        {
-          id: "issuing_inspector_name_label",
-          type: "LABEL",
-          required: false,
-          "context-value": "Sincerely,",
-          "read-only": false,
-        },
-        {
-          id: "issuing_inspector_name",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_email",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_phone",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_email",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_phone_number",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_fax_number",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_mailing_address_line_1",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "rc_office_mailing_address_line_2",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_type_code",
-          required: false,
-          "read-only": true,
-        },
-      ],
-    },
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "SSF",
-    description: "Submitted Shape Files",
-    now_application_document_sub_type_code: "MDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "CSL",
-    description: "Cross-sectional/Longitudinal",
-    now_application_document_sub_type_code: "MDO",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "NTR",
-    description: "Notice of Work Form",
-    now_application_document_sub_type_code: "AEF",
-    document_template: {
-      document_template_code: "NTR",
-      form_spec: [],
-    },
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "PMT",
-    description: "Working Permit",
-    now_application_document_sub_type_code: "AEF",
-    document_template: {
-      document_template_code: "PMT",
-      form_spec: [
-        {
-          id: "mine_no",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "permittee",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "property",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_name",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_date",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_type",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_type_code",
-          required: false,
-          "read-only": true,
-        },
-      ],
-    },
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "PMA",
-    description: "Working Permit for Amendment",
-    now_application_document_sub_type_code: "AEF",
-    document_template: {
-      document_template_code: "PMA",
-      form_spec: [
-        {
-          id: "mine_no",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "permittee",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "property",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "issuing_inspector_name",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_date",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_type",
-          required: false,
-          "read-only": true,
-        },
-        {
-          id: "application_type_code",
-          required: false,
-          "read-only": true,
-        },
-      ],
-    },
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "ANS",
-    description: "Annual Summary",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "ACP",
-    description: "Archaeological Chance Find Procedure",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "BLP",
-    description: "Blasting Procedure",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "EMS",
-    description: "Explosives Magazine Storage and Use Permit Application",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "LAL",
-    description: "Landowner Authorization Letter",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "MRP",
-    description: "Mine Emergency Response Plan",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "RFE",
-    description: "Record of First Nations Engagement",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "TAL",
-    description: "Tenure Authorization Letter",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "AOA",
-    description: "Archaeological Overview Assessment",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-  {
-    now_application_document_type_code: "AIA",
-    description: "Archaeological Impact Assessment",
-    now_application_document_sub_type_code: "AAF",
-    document_template: {},
-    active_ind: true,
-  },
-];
-
-export const DROPDOWN_APPLICATION_DOCUMENT_TYPES = [
-  {
-    value: "WFI",
-    label: "30 day Warning for Information",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "WFS",
-    label: "30 day Warning for Security",
-    isActive: true,
-    subType: "SDO",
-  },
-  {
-    value: "CAL",
-    label: "Acknowledgement Letter",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "AKL",
-    label: "Acknowledgement of Security Letter",
-    isActive: true,
-    subType: "SDO",
-  },
-  {
-    value: "PCA",
-    label: "Advertisement",
-    isActive: true,
-    subType: "PDO",
-  },
-  {
-    value: "ARE",
-    label: "Agent Letter of Representation",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "AMR",
-    label: "Amendment Request",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "ANS",
-    label: "Annual Summary",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "ACP",
-    label: "Archaeological Chance Find Procedure",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "AIA",
-    label: "Archaeological Impact Assessment",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "AOA",
-    label: "Archaeological Overview Assessment",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "BCR",
-    label: "Begin Consultation Request",
-    isActive: true,
-    subType: "CDO",
-  },
-  {
-    value: "BRR",
-    label: "Begin Referral Request",
-    isActive: true,
-    subType: "RDO",
-  },
-  {
-    value: "BLP",
-    label: "Blasting Procedure",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "SCD",
-    label: "Bond Calculator",
-    isActive: true,
-    subType: "SDO",
-  },
-  {
-    value: "CCC",
-    label: "Consultation Correspondence (not in CRTS)",
-    isActive: true,
-    subType: "CDO",
-  },
-  {
-    value: "CRS",
-    label: "Consultation Report/Summary",
-    isActive: true,
-    subType: "CDO",
-  },
-  {
-    value: "CSD",
-    label: "Consultation Support for Decision",
-    isActive: true,
-    subType: "CDO",
-  },
-  {
-    value: "CSL",
-    label: "Cross-sectional/Longitudinal",
-    isActive: true,
-    subType: "MDO",
-  },
-  {
-    value: "DWP",
-    label: "Description of Work/Work Program",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "ECC",
-    label: "Email Correspondence/Communications",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "EMS",
-    label: "Explosives Magazine Storage and Use Permit Application",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "FDP",
-    label: "Fugitive Dust Management Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "LTM",
-    label: "Land Title/Licence of Occupation Map",
-    isActive: true,
-    subType: "MDO",
-  },
-  {
-    value: "LAL",
-    label: "Landowner Authorization Letter",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "LNO",
-    label: "Landowner Notification",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "LMA",
-    label: "Location Map",
-    isActive: true,
-    subType: "MDO",
-  },
-  {
-    value: "MYA",
-    label: "MYAB Update Form",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "MAD",
-    label: "Metal Leaching/Acid Rock Drainage",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "MRP",
-    label: "Mine Emergency Response Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "PCM",
-    label: "Ministry Response",
-    isActive: true,
-    subType: "PDO",
-  },
-  {
-    value: "NIA",
-    label: "No Interest Acknowledgement Form",
-    isActive: true,
-    subType: "SDO",
-  },
-  {
-    value: "NPR",
-    label: "No Permit Required",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "NPI",
-    label: "No Permit Required IP",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "NTR",
-    label: "Notice of Work Form",
-    isActive: true,
-    subType: "AEF",
-  },
-  {
-    value: "OTH",
-    label: "Other",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "OTA",
-    label: "Other",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "OMP",
-    label: "Other Management Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "OMA",
-    label: "Overview Map",
-    isActive: true,
-    subType: "MDO",
-  },
-  {
-    value: "NPE",
-    label: "Permit Enclosed Letter",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "PFR",
-    label: "Preliminary Field Reconnaissance",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "MPW",
-    label: "Proposed and/or Permitted Mine Area Map",
-    isActive: true,
-    subType: "MDO",
-  },
-  {
-    value: "MPG",
-    label: "Proposed and/or Permitted Mine Area Map",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "PCC",
-    label: "Public Comment",
-    isActive: true,
-    subType: "PDO",
-  },
-  {
-    value: "RFD",
-    label: "Reasons for Decision",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "RPL",
-    label: "Reclamation Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "RFE",
-    label: "Record of First Nations Engagement",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "RLE",
-    label: "Referral Letter (outside of E-Referral)",
-    isActive: true,
-    subType: "RDO",
-  },
-  {
-    value: "RRE",
-    label: "Referral Response (outside of E-Referral)",
-    isActive: true,
-    subType: "RDO",
-  },
-  {
-    value: "RSR",
-    label: "Referral Summary Roll Up",
-    isActive: true,
-    subType: "RDO",
-  },
-  {
-    value: "RJL",
-    label: "Rejection Letter",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "RMI",
-    label: "Request for More Information",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "REV",
-    label: "Review",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "RSP",
-    label: "Riparian Setbacks Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "SRB",
-    label: "Scan of Reclamation Security Document",
-    isActive: true,
-    subType: "SDO",
-  },
-  {
-    value: "SEP",
-    label: "Sediment and Erosion Control Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "SOP",
-    label: "Standard/Safe Operating Procedures",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "SRE",
-    label: "Status Report",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "SOM",
-    label: "Status Report - Overlapping Interests Maps",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "SRS",
-    label: "Status Report - Shape Files",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "SSF",
-    label: "Submitted Shape Files",
-    isActive: true,
-    subType: "MDO",
-  },
-  {
-    value: "SMA",
-    label: "Supplemental Map",
-    isActive: true,
-    subType: "MDO",
-  },
-  {
-    value: "SUD",
-    label: "Supporting Documents",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "TAL",
-    label: "Tenure Authorization Letter",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "TSS",
-    label: "Terrain Stability Study",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "TMP",
-    label: "Title/Tenure Map",
-    isActive: true,
-    subType: "MDO",
-  },
-  {
-    value: "VMP",
-    label: "Vegetation Management Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "WMP",
-    label: "Water Management Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "WPL",
-    label: "Wildlife Management Plan",
-    isActive: true,
-    subType: "AAF",
-  },
-  {
-    value: "WDL",
-    label: "Withdrawal Letter",
-    isActive: true,
-    subType: "GDO",
-  },
-  {
-    value: "PMT",
-    label: "Working Permit",
-    isActive: true,
-    subType: "AEF",
-  },
-  {
-    value: "PMA",
-    label: "Working Permit for Amendment",
-    isActive: true,
-    subType: "AEF",
-  },
-];
-
-export const APPLICATION_DOCUMENT_TYPES_HASH = {
-  PMA: "Working Permit for Amendment",
-  PMT: "Working Permit",
-  WDL: "Withdrawal Letter",
-  WPL: "Wildlife Management Plan",
-  WMP: "Water Management Plan",
-  VMP: "Vegetation Management Plan",
-  TMP: "Title/Tenure Map",
-  TSS: "Terrain Stability Study",
-  TAL: "Tenure Authorization Letter",
-  SUD: "Supporting Documents",
-  SMA: "Supplemental Map",
-  SSF: "Submitted Shape Files",
-  SRS: "Status Report - Shape Files",
-  SOM: "Status Report - Overlapping Interests Maps",
-  SRE: "Status Report",
-  SOP: "Standard/Safe Operating Procedures",
-  SEP: "Sediment and Erosion Control Plan",
-  SRB: "Scan of Reclamation Security Document",
-  RSP: "Riparian Setbacks Plan",
-  REV: "Review",
-  RMI: "Request for More Information",
-  RJL: "Rejection Letter",
-  RSR: "Referral Summary Roll Up",
-  RRE: "Referral Response (outside of E-Referral)",
-  RLE: "Referral Letter (outside of E-Referral)",
-  RFE: "Record of First Nations Engagement",
-  RPL: "Reclamation Plan",
-  RFD: "Reasons for Decision",
-  PCC: "Public Comment",
-  MPG: "Proposed and/or Permitted Mine Area Map",
-  MPW: "Proposed and/or Permitted Mine Area Map",
-  PFR: "Preliminary Field Reconnaissance",
-  NPE: "Permit Enclosed Letter",
-  OMA: "Overview Map",
-  OMP: "Other Management Plan",
-  OTA: "Other",
-  OTH: "Other",
-  NTR: "Notice of Work Form",
-  NPI: "No Permit Required IP",
-  NPR: "No Permit Required",
-  NIA: "No Interest Acknowledgement Form",
-  PCM: "Ministry Response",
-  MRP: "Mine Emergency Response Plan",
-  MAD: "Metal Leaching/Acid Rock Drainage",
-  MYA: "MYAB Update Form",
-  LMA: "Location Map",
-  LNO: "Landowner Notification",
-  LAL: "Landowner Authorization Letter",
-  LTM: "Land Title/Licence of Occupation Map",
-  FDP: "Fugitive Dust Management Plan",
-  EMS: "Explosives Magazine Storage and Use Permit Application",
-  ECC: "Email Correspondence/Communications",
-  DWP: "Description of Work/Work Program",
-  CSL: "Cross-sectional/Longitudinal",
-  CSD: "Consultation Support for Decision",
-  CRS: "Consultation Report/Summary",
-  CCC: "Consultation Correspondence (not in CRTS)",
-  SCD: "Bond Calculator",
-  BLP: "Blasting Procedure",
-  BRR: "Begin Referral Request",
-  BCR: "Begin Consultation Request",
-  AOA: "Archaeological Overview Assessment",
-  AIA: "Archaeological Impact Assessment",
-  ACP: "Archaeological Chance Find Procedure",
-  ANS: "Annual Summary",
-  AMR: "Amendment Request",
-  ARE: "Agent Letter of Representation",
-  PCA: "Advertisement",
-  AKL: "Acknowledgement of Security Letter",
-  CAL: "Acknowledgement Letter",
-  WFS: "30 day Warning for Security",
-  WFI: "30 day Warning for Information",
-};
-
-export const APPLICATION_PROGRESS_STATUS_CODES = {
-  records: [
-    {
-      application_progress_status_code: "VER",
-      description: "Verification",
-    },
-    {
-      application_progress_status_code: "REV",
-      description: "Technical Review",
-    },
-    {
-      application_progress_status_code: "REF",
-      description: "Referral / Consultation",
-    },
-    {
-      application_progress_status_code: "DEC",
-      description: "Decision",
-    },
-  ],
-};
-
-export const APPLICATION_PERMIT_TYPES = {
-  records: [
-    {
-      now_application_permit_type_code: "MY-ABP",
-      description: "Multi-Year, Area-Based Permit",
-    },
-    {
-      now_application_permit_type_code: "OYP",
-      description: "One-Year Permit",
-    },
-    {
-      now_application_permit_type_code: "MYP",
-      description: "Multi-Year Permit",
-    },
-  ],
-};
-
-export const DROPDOWN_APPLICATION_PERMIT_TYPES = [
-  { value: "MYP", label: "Multi-Year Permit", subType: null, isActive: true },
-  {
-    value: "MY-ABP",
-    label: "Multi-Year, Area-Based Permit",
-    subType: null,
-    isActive: true,
-  },
-  { value: "OYP", label: "One-Year Permit", subType: null, isActive: true },
-];
-
-export const NOTICE_OF_WORK_REVIEW_RESPONSE = {
-  records: [
-    {
-      now_application_review_id: 2,
-      now_application_guid: "587507c0-cafd-43d5-ae40-26934e039893",
-      now_application_review_type_code: "FNC",
-      response_date: "2019-12-16",
-      referee_name: "William Davila",
-      documents: [
-        {
-          now_application_document_xref_guid: "35a7c16b-32d6-4265-93c4-304e9b3e2e8c",
-          now_application_document_type_code: "REV",
-          now_application_document_type_code_description: "Review",
-          description: null,
-          is_final_package: null,
-          mine_document: {
-            mine_document_guid: "8157b615-4522-4121-ab02-0ba1f1e0a4b4",
-            mine_guid: "f66e6fdc-9bad-4cf0-baae-6665721ca4ba",
-            document_manager_guid: "338ab757-572c-40ec-be71-c9620daad86e",
-            document_name: "0101581201901_APPLICATION_FORM (1).pdf",
-            upload_date: "2020-01-13T23:50:06.287515+00:00",
-          },
-        },
-        {
-          now_application_document_xref_guid: "bede5549-e8b4-40c7-bb73-abb3bc41098d",
-          now_application_document_type_code: "REV",
-          now_application_document_type_code_description: "Review",
-          description: null,
-          is_final_package: null,
-          mine_document: {
-            mine_document_guid: "eeeb6f1e-9f34-490f-a240-947f2a59eb0f",
-            mine_guid: "f66e6fdc-9bad-4cf0-baae-6665721ca4ba",
-            document_manager_guid: "6749bf0c-4640-487e-8cde-0b9ac972e214",
-            document_name: "0101581201901_APPLICATION_FORM (2).pdf",
-            upload_date: "2020-01-13T23:50:06.289240+00:00",
-          },
-        },
-      ],
-    },
-    {
-      now_application_review_id: 7,
-      now_application_guid: "587507c0-cafd-43d5-ae40-26934e039893",
-      now_application_review_type_code: "FNC",
-      response_date: "2020-01-01",
-      referee_name: "asdasdwdq",
-      documents: [
-        {
-          now_application_document_xref_guid: "a9e83a40-9030-4c36-a2b8-b5897123083a",
-          now_application_document_type_code: "REV",
-          now_application_document_type_code_description: "Review",
-          description: null,
-          is_final_package: null,
-          mine_document: {
-            mine_document_guid: "5cf78f22-4d7d-4cd6-8ce8-80e01ed71e1a",
-            mine_guid: "f66e6fdc-9bad-4cf0-baae-6665721ca4ba",
-            document_manager_guid: "d98d09e1-2780-4e32-b62d-b5f496c098af",
-            document_name: "0101581201901_APPLICATION_FORM.pdf",
-            upload_date: "2020-01-13T23:50:27.039598+00:00",
-          },
-        },
-      ],
-    },
-  ],
-};
-
-export const GENERATABLE_APPLICATION_DOCUMENTS = {
-  RJL: {
-    now_application_document_type_code: "RJL",
-    description: "Rejection Letter",
-    document_template: {
-      document_template_code: "NOW-RJL",
-      form_spec: [
-        { id: "letter_dt", label: "Letter Date", type: "DATE", placeholder: null, required: true },
-        {
-          id: "mine_no",
-          label: "Mine Number",
-          type: "FIELD",
-          placeholder: "Enter the mine number",
-          required: true,
-        },
-        {
-          id: "proponent_address",
-          label: "Proponent Address",
-          type: "FIELD",
-          placeholder: "Enter the proponent's address",
-          required: true,
-        },
-        {
-          id: "proponent_name",
-          label: "Proponent Name",
-          type: "FIELD",
-          placeholder: "Enter the proponent's name",
-          required: false,
-        },
-        {
-          id: "property",
-          label: "Property",
-          type: "FIELD",
-          placeholder: "Enter the property",
-          required: true,
-        },
-        {
-          id: "apl_date",
-          label: "Application Date",
-          type: "DATE",
-          placeholder: null,
-          required: true,
-        },
-        {
-          id: "inspector",
-          label: "Inspector",
-          type: "FIELD",
-          placeholder: "Enter the inspector's name",
-          required: true,
-        },
-      ],
-    },
-  },
-  WDL: {
-    now_application_document_type_code: "WDL",
-    description: "Withdrawal Letter",
-    document_template: {
-      document_template_code: "NOW-WDL",
-      form_spec: [
-        { id: "letter_dt", label: "Letter Date", type: "DATE", placeholder: null, required: true },
-        {
-          id: "file_no",
-          label: "File Number",
-          type: "FIELD",
-          placeholder: "Enter the file number",
-          required: true,
-        },
-        {
-          id: "proponent_address",
-          label: "Proponent Address",
-          type: "FIELD",
-          placeholder: "Enter the proponent's address",
-          required: true,
-        },
-        {
-          id: "proponent_name",
-          label: "Proponent Name",
-          type: "FIELD",
-          placeholder: "Enter the proponent's name",
-          required: false,
-        },
-        {
-          id: "property",
-          label: "Property",
-          type: "FIELD",
-          placeholder: "Enter the property",
-          required: true,
-        },
-        {
-          id: "withdrawal_dt",
-          label: "Withdrawal Date",
-          type: "DATE",
-          placeholder: null,
-          required: true,
-        },
-        {
-          id: "inspector",
-          label: "Inspector",
-          type: "FIELD",
-          placeholder: "Enter the inspector's name",
-          required: true,
-        },
-      ],
-    },
-  },
-  CAL: {
-    now_application_document_type_code: "CAL",
-    description: "Acknowledgement Letter",
-    document_template: {
-      document_template_code: "NOW-CAL",
-      form_spec: [
-        { id: "letter_dt", label: "Letter Date", type: "DATE", placeholder: null, required: true },
-        {
-          id: "mine_no",
-          label: "Mine Number",
-          type: "FIELD",
-          placeholder: "Enter the mine number",
-          required: true,
-        },
-        {
-          id: "proponent_address",
-          label: "Proponent Address",
-          type: "FIELD",
-          placeholder: "Enter the proponent's address",
-          required: true,
-        },
-        {
-          id: "proponent_name",
-          label: "Proponent Name",
-          type: "FIELD",
-          placeholder: "Enter the proponent's name",
-          required: false,
-        },
-        {
-          id: "emailed_to",
-          label: "Emailed to",
-          type: "FIELD",
-          placeholder: "Enter the name of the email recipient",
-          required: false,
-        },
-        {
-          id: "property",
-          label: "Property",
-          type: "FIELD",
-          placeholder: "Enter the property",
-          required: true,
-        },
-        {
-          id: "apl_dt",
-          label: "Application Date",
-          type: "DATE",
-          placeholder: null,
-          required: true,
-        },
-        {
-          id: "exploration_type",
-          label: "Exploration Type",
-          type: "FIELD",
-          placeholder: "Enter the exploration type",
-          required: true,
-        },
-        {
-          id: "bond_inc_amt",
-          label: "Bond Amount",
-          type: "FIELD",
-          placeholder: "Enter the bond amount",
-          required: true,
-        },
-        {
-          id: "inspector",
-          label: "Inspector",
-          type: "FIELD",
-          placeholder: "Enter the inspector's name",
-          required: true,
-        },
-      ],
-    },
-  },
-};
-
-export const NOW_APPLICATION_DELAY = [
-  {
-    now_application_delay_guid: "c68dfefb-af11-48c5-8d36-39704b8385fc",
-    delay_type_code: "OAB",
-    start_comment: "mock delay",
-    start_date: "2021-10-26T18:05:19.380715+00:00",
-    end_comment: "mock stopping coment",
-    end_date: "2021-10-26T18:05:56.475794+00:00",
-  },
-];
